+++ tools/generator/domain/src/main/java/is/codion/tools/generator/domain/DomainSource.java
@@ -727,7 +727,7 @@ public final class DomainSource {
 	private boolean identityKeyGeneratorUsed() {
 		return sortedDefinitions.stream()
 						.map(entityDefinition -> entityDefinition.primaryKey().generator())
-						.anyMatch(KeyGenerator.Identity.class::isInstance);
+						.anyMatch(generator -> generator == KeyGenerator.identity());
 	}
 
 	static String underscoreToCamelCase(String text) {